stable v1.0

+++ app/src/main/java/com/ho/elhawarycenter/model/Case.kt
@@ -12,7 +12,7 @@ import kotlinx.parcelize.Parcelize
 data class Case(
     @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "id") val id: Int,
     @ColumnInfo(name = "name") val name: String,
-    @ColumnInfo(name = "age") val age: Int,
+    @ColumnInfo(name = "age") val age: Float,
     @ColumnInfo(name = "diagnosis") val diagnosis: String,
     @ColumnInfo(name = "type") val type: String,
     @ColumnInfo(name = "gender") val gender: String,
